Id 형변환 (Integer to Long)에 따른 대응 수정 등

+++ src/main/java/pinup/backend/notice/command/service/NoticeCommandService.java
@@ -18,8 +18,8 @@ public class NoticeCommandService {
     private final NoticeRepository noticeRepository;
     private final AdminRepository adminRepository;
 
-    public Integer postNotice(NoticePostRequest request) {
-        Admin admin = adminRepository.findById((int) request.getAdminId().longValue()).orElseThrow(IllegalArgumentException::new);
+    public Long postNotice(NoticePostRequest request) {
+        Admin admin = adminRepository.findById(request.getAdminId()).orElseThrow(IllegalArgumentException::new);
 
         return noticeRepository.save(Notice.builder()
                 .noticeContent(request.getNoticeContent())
